requirements.txt
django >=2, <3
django-tinymce
